Refactor to separate go files.
